fix: home 화면 변경 작업 완료
